Testing out BLE
I was going to use the RNBT Serial package to send commands, but it has not been supported for a long while. I am going to try to make the BLE module work as it is actively supported. Installed module, ran their test example, began modifying Settings to include BLE and permissions. Also renamed Home export from APP to HOME.