Heavy reconfig of onPressLamp function
Function now is much smaller, only containing a single ble write. What is written is dynamic based on settings and lamp pressed. Also reconfigured how state colors variable is laid outand renamed it to lamps. This was done to allow reference to dynamic names of lamps in the lamps objects (ie lamps[lampname]